+++ spec/rubocop/cop/lint/useless_assignment_spec.rb
@@ -1359,6 +1359,23 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment, :config do
     end
   end
 
+  context 'when variables are assigned using unary operator in chained assignment and remain unreferenced' do
+    it 'registers an offense' do
+      expect_offense(<<~RUBY)
+        def some_method
+          foo = -bar = do_something
+          ^^^ Useless assignment to variable - `foo`.
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        def some_method
+          -bar = do_something
+        end
+      RUBY
+    end
+  end
+
   context 'when variables are assigned with sequential assignment using the comma operator and unreferenced' do
     it 'registers an offense' do
       expect_offense(<<~RUBY)
@@ -1423,12 +1440,48 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment, :config do
     end
   end
 
+  context 'when part of a multiple assignment is enclosed in parentheses with splat' do
+    it 'registers an offense when the variable is not used' do
+      expect_offense(<<~RUBY)
+        def some_method
+          (foo, bar), *baz = do_something
+                       ^^^ Useless assignment to variable - `baz`. Use `_` or `_baz` as a variable name to indicate that it won't be used.
+          puts foo, bar
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        def some_method
+          (foo, bar), *_ = do_something
+          puts foo, bar
+        end
+      RUBY
+    end
+
+    it 'registers an offense when the variable is not used in nested assignment' do
+      expect_offense(<<~RUBY)
+        def some_method
+          foo, (*bar, baz) = do_something
+                 ^^^ Useless assignment to variable - `bar`. Use `_` or `_bar` as a variable name to indicate that it won't be used.
+          puts foo, baz
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        def some_method
+          foo, (*_, baz) = do_something
+          puts foo, baz
+        end
+      RUBY
+    end
+  end
+
   context 'when a variable is assigned with rest assignment and unreferenced' do
     it 'registers an offense' do
       expect_offense(<<~RUBY)
         def some_method
           foo, *bar = do_something
-                ^^^ Useless assignment to variable - `bar`.
+                ^^^ Useless assignment to variable - `bar`. Use `_` or `_bar` as a variable name to indicate that it won't be used.
           puts foo
         end
       RUBY